Update CloudFlare codec, resolves #9 (#10)
* # If applied, this commit will...
Correct handling of OriginResponseTimeMillis field in CloudFlare codec

# Explain why this change is being made
Users reported that the Lamba was failing with uncaught exceptions.  These exceptions turned out to be ClassCastExceptions related to the handling of OriginResponseTimeMillis.  The exceptions were not manifesting in JUnit tests, but did appear when the code was executed as an AWS Lambda.  The change also enables stack trace logging for uncaught exceptions.

# Provide links to any relevant tickets, articles or other resources
Related Tickets:
https://support.graylog.com/#ticket/zoom/3167

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch cloudflare-lambda-ticket
# Changes to be committed:
#       modified:   README.md
#       modified:   src/main/java/org/graylog/integrations/s3/GraylogS3Function.java
#       modified:   src/main/java/org/graylog/integrations/s3/codec/CloudflareLogCodec.java
#       modified:   src/test/java/org/graylog/integrations/s3/ContentTypeTest.java
#

* # If applied, this commit will...
Update JUnit test to actually test Unix nano timestamp

# Explain why this change is being made
CloudFlareLogsParserTest did not previously check with an actual unix nano timestamp.  Updated the field in UNIX_NANO_TIMESTAMP_MESSAGE and verified that the test now fails on master but passes on the branch with the fix.

# Provide links to any relevant tickets, articles or other resources
Related Issues:
https://github.com/Graylog2/graylog-s3-lambda/issues/9